eliminar

+++ mvc/vista/ver_producto.php
@@ -4,6 +4,7 @@ include_once("../modelo/Producto.php");
 
 	$controlador = new Producto();
 	$sql= $controlador->listarProducto();
+	$bandera="false";
 
  ?> 
 
@@ -23,24 +24,55 @@ include_once("../modelo/Producto.php");
  		<th>ProductoFechaAltaDB</th>
  		<th>ProductoFechaBajaDB</th>
  		<th>ProductoEstado</th>
+ 		<th><a <form id="bandera" method="post" action="<?php $bandera="true"; ?>"> 
+ 		      <input id="button" type="button" onClick="document.getElementById('frmcancel').submit()" value="Ver Inactivos">
+ 		      </form></a></th>
+
+ 		
  	</thead>
  	<tbody>
  <?php  
- 		foreach($sql as $row){ ?>
- 		<tr>
- 			<td><?php echo "{$row['IDProducto']}"; ?></td>
- 			<td><?php echo "{$row['ProductoNombre']}"; ?></td>
- 			<td><?php echo "{$row['ProductoPrecio']}"; ?></td>
- 			<td><?php echo "{$row['ProductoFechaAltaDB']}"; ?></td>
- 			<td><?php echo "{$row['ProductoFechaBajaDB']}"; ?></td>
- 			<td><?php echo "{$row['ProductoEstado']}"; ?></td>
- 			<?php /* <td><a href="?cargar=editar_producto&IDProducto=<?php echo $row['IDProducto'] ?>"> Modificar</a></td>
- 			*/?>
- 			<td><a href="editar_producto.php?IDProducto=<?php echo $row['IDProducto'] ?>"> Modificar Producto</a></td>
- 			
- 			<td><a href="eliminar_producto.php?IDProducto=<?php echo $row['IDProducto'] ?>"> Eliminar Producto</a></td>
+ 		foreach($sql as $row){ 
+ 			if ($bandera=="false") {
+
+ 				if ($row['ProductoEstado']=="Activo") { ?>
+ 				<tr>
+	 			<td><?php echo "{$row['IDProducto']}"; ?></td>
+	 			<td><?php echo "{$row['ProductoNombre']}"; ?></td>
+	 			<td><?php echo "{$row['ProductoPrecio']}"; ?></td>
+	 			<td><?php echo "{$row['ProductoFechaAltaDB']}"; ?></td>
+	 			<td><?php echo "{$row['ProductoFechaBajaDB']}"; ?></td>
+	 			<td><?php echo "{$row['ProductoEstado']}"; ?></td>
+
+	 			<td><a href="editar_producto.php?IDProducto=<?php echo $row['IDProducto'] ?>"> Modificar Producto</a></td>
+	 			
+	 			<td><a href="eliminar_producto.php?IDProducto=<?php echo $row['IDProducto'] ?>"> Eliminar Producto</a></td>
 
  		</tr>
+
+ 		<?php
+ 			}
+ 			}else{ ?>
+ 			<tr>
+	 			<td><?php echo "{$row['IDProducto']}"; ?></td>
+	 			<td><?php echo "{$row['ProductoNombre']}"; ?></td>
+	 			<td><?php echo "{$row['ProductoPrecio']}"; ?></td>
+	 			<td><?php echo "{$row['ProductoFechaAltaDB']}"; ?></td>
+	 			<td><?php echo "{$row['ProductoFechaBajaDB']}"; ?></td>
+	 			<td><?php echo "{$row['ProductoEstado']}"; ?></td>
+
+	 			<td><a href="editar_producto.php?IDProducto=<?php echo $row['IDProducto'] ?>"> Modificar Producto</a></td>
+	 			
+	 			<td><a href="eliminar_producto.php?IDProducto=<?php echo $row['IDProducto'] ?>"> Eliminar Producto</a></td>
+
+ 		</tr>
+ 		<?php
+ 		}
+ 			?>
+ 			
+ 			
+
+ 		
 <?php } ?>
 			</tbody>
  </table>
